!! IT FINALLY COMPILES !!
needs a LOT of work in terms of optimization.
the 64-bit div instruction impl is entirely missing (why on earth is there not a stdlib function for that?).
predefined assembler symbols are not being predefined (linker is fine though).

where c# lets you easily iterate through an enum and conert values to strings (used in the c# version's syscall/errcode predefines), this is not present in c++.
options are: make a table to read them from internally (might be able to make it a constexpr system which would be nice and efficient), or read them from a file.
so far, leaning toward the file route as it'd add a layer of modularity for the user at the (small) expense of speed.

this commit only marks that it compiled - it's still completely untested...