major bug fixes + asm predefs
fixed a bug where TryAppendExpr() was getting segment position **after** writing the dummy, which would guarantee an out of bounds on hole patching immediately after.
(in the c# source this was correct, not sure why/when i changed the order during translation).

fixed a (minor) bug where the (error) line number was being duplicated in the PatchError::Error case in TryAppendExpr().
it puts the line of the hole so that the linker can tell you where the error occurred in the file, but TryAppendExpr() is during assemble time and thus already has that info.

fixed a bug in RenameSymbol() where passing a value sourced from Symbols/GlobalSymbols/ExternalSymbols could result in using dangling references if that symbol was what was removed.
function signature now takes std::string by value.
removed a leftover `static` declarator in the impl file.
optimized out a set lookup by using the `expr` pointer returned from TryGetValue().

added the assembler predefines (only for sys_ and err_ - others from the C# impl are intentionally left out for now - esp. __time__).

changed the elim_symbols and rename_symbols vectors to store std::string by value (ensures a removal/insertion doesn't invalidate the pointers).

fixed a bug where ObjectFile::WriteTo() and ReadFrom() were using the &vec[0] idiom (which fails if the vector is empty).
these now directly use std::vector::data().
in the case that the vector is empty, it's ok because the length to write is zero.

fixed a bug where the cmd line arg parser was using `argc` as the length of `argv[i]` for handling shortnames.
now directly controls a character pointer and tests for null - this also eliminates the need for a temp char* to remember what the working arg is in the case of a -o flag.
removed a break statement leftover from the C# impl that was causing only the first arg to be processed.

changed pathspec to hold std::string instead of const char*.
paths are now inserted via emplace_back() to avoid a move ctor.