Computer class cleanup
worked on cleaning up the Computer class's interface.
hid all the data that was marked as public for convenience during testing.

added the ST_Wraper class (nested in Computer) to make FPU access fancier.
acts as a reference to a long double and internally handles updating the tag word.
additionally, has static definitions for the tag meanings, a method to get the referenced ST register's tag, and convenience functions for testing the specific tag values.
also has operator<< overloaded for std::ostream - writes "Empty" if empty, otherwise puts the long double value it wraps.
has a Free() method to mark that ST register as empty.

increased default max memory size to 8GB (apparently overflow at compile-time is saturated to zero, so that's fixed now as well).

marked use of aligned_malloc and aligned_free (from Utility.h) with explicit CSX64:: (because those are fairly common symbol names, but not part of the C++ standard).

fixed an indenting bug, which is gonna make this commit have a ton of +/- :(

fixed a few occurrences of memory setters not checking the readonly barrier.

realized there was no external access for ZMM registers, so added ZMM(int) - returns a reference.

changed Syscall() to ProcessSYSCALL to be more like all the other opcode handlers.

added ctor/dtor/asgn to FileDescriptor:
is default-constructible as empty.
can be moved - cannot be copied.
this showed a bug in Process_sys_lseek() where the FileDescriptor was being copied.

added ctor/dtor/asgn to Computer:
is default-constructible as per usual
cannot be moved or copied.
added a friend swap() function for ADL lookup.

the Computer class is large enough that even a move would be fairly expensive.
if this behavior is desired, it would be better to hold a (smart) pointer to it instead of the actual object.
since Tick() can perform 2^64 operations in one function call, this wouldn't even hurt execution performance.