64-bit divide impl
added Neg_128(u64&, u64&) - handles extended negatives more conveniently.

added impl for signed/unsigned divide.
signed is done in terms of unsigned (if xor of sign flags is 1, quotient *and remainder* are negated).
in the case where the high bits are zero, uses built-in operators for speed. otherwise uses the unsigned binary long division algorithm extended to 2 64-bit values.

with this now done, the C++ impl is finally caught up to the C# impl.