added simd fp comparison
added a ton of simd fp comparison instructions - ridiculous number of assembler mnemonics, but all map to the VPUBinary format, with an 8-bit extension code for the comparison to use.
changed the VPUBinary formatter to allow extended op codes.

added TryProcessVPU_FCMP() - takes the last arg as an instant imm comparison predicate, pops it off, and hands it over to VPUBinary as the extended op code.

added a ton of handlers for all the various instruction mnemonics during assemble-time.
CMPSD (string) now conflicts with CMPSD (vec) - moved into the disambig section.

for execution, each comparison predicate has its own handler, which is generated via macro from a tabular reference.
the handlers are organized in a lookup table for speed.

removed a bunch of occurrences of multiple spaces - resulted in a lot of +-.

fixed a bug where the random generator object wasn't getting a mutable seed (apparently the default ctor is a constant seed).
now initialized with the result of (unsigned int)std::time(nullptr).