added vec sqrt/rsqrt
added the VPUUnary format - just like the binary version but src2 is the only source (no src1 register).
takes a VPUUnaryDelegate, which is just like the binary one except it has 1 fewer u64 inputs.

used this format to implement vec instructions for sqrt and rsqrt.

auto-replaced use of "GetMemAdv(1, " with "GetMemAdv<u8>(" for efficiency.
changed a few uses of "(blah & something) != 0" to "blah & something" - this is the default behavior in C++ and makes it cleaner (was desired in the C# impl, but C# doesn't support it).