added COMISD/COMISS + bug fixes
added TryProcessVPUBinary_2arg() - as TryProcessVPUBinary() but forces the 2 operand case.

used this helper to implement the new COMISD and COMISS instructions.
as explained in the comments above the execution code, these use the VPUBinary format to avoid having to make a whole new (and complex) vpu format just for one instruction.
the assembler ensures everything works out so that it behaves in the standard manner, but the user could potentially write weirdness in machine code (because it can modify the destination despite being a comparison).

fixed a bug where Expr::__Evaluate__() wasn't catching division by zero.
would make the assembler/linker silently fail (apparently div-by-zero in C++ doesn't throw an exception - it just calls exit()).
now properly yields an asembler/link error.
with how the current imm system works, unless it's a critical expression, the error will be put off until link time, which is a bit weird: i.e. "1/0" is a link error unless it's used as a critical expression, in which case it's an assemble error.