bug fix + more unique_ptr
fixed a bug where Ptrdiff() was using VerifyLegalExpressionIgnores for finding pointer bases to eliminate.
this resulted in pointer differences not being processed as instant imms.
was introduced in the C# impl prior to starting the C++ translation and wasn't found because, while pointer differences were used in the test files, none were used in critical expressions.
new unordered set PtrdiffIDs contains the strings to use for matching - fixing the problem.

changed all the imm parser backend functions to use unique_ptr.

PopulateAddSub() is only used in Ptrdiff().
because of this, i've removed the current overloads.
replaced with a specialized form for new Ptrdiff() code, which takes vector<Expr> and is performed on rvalues for added efficiency.

ChainAddition() likewise now takes vector<Expr> instead of vector<const Expr*>.