eschew new/delete
all uses of new/delete are now gone (aside from in FileDescriptor::Close(), which may be requested to manage the stream internally by the Open() function with managed=true).

added back in the assembler predefines from the C# impl (forgot to go back and do those).
the __version__ predefine isn't in yet, as this would create a dependency on Computer.h, which i don't want to do.
the current plan is to create separate version numbers for the execution engine and the assembler.
__version__ would then refer to the assembly version and there would be some instruction or syscall to get the execution engine version number at runtime.

added static_assert statements to CoreTypes.h to ensure assumptions about the sizes of floating-point types are sound.
added typedefs f32 and f64 to CoreTypes.h, which are safeguarded by assert statements.

removed a big block comment for a .ToString() method from Expr - was left in as pseudocode for writing the operator<< for Expr, but was never removed.

added an extra step to the sys_open() algorithm to ensure the stream object can't leak if fd->Open() fails.