added fsave/frstor and fstenv/fldenv
fixed an "incorrect" error message in TryProcessFSTLD_WORD() - if _sizecode wasn't matched, said there was a mismatch, but _sizecode is supplied by the program, not by the user, which would be a bit confusing.
new error message just says the size is not supported.
added a comment this function in the header that explains that _sizecode is only used for explicit memory operand size validation.

added fsave/frstor and fstenv/fldenv - internally, use extra modes of FSTLD_WORD.
fsave/frstor write the ST registers with native f64 instead of the f80 it's technically supposed to use.
converting to twords would be slow and pointless since it would be a software conversion and there aren't any CSX64 instructions that handle twords anyway.
the handlers for these instructions passes 0xff..ff to TryProcessFSTLD_WORD() to ensure the user can't specify an explicit size (because they aren't powers of 2).

changed the static_assert code for ensuring double and float are 64 and 32-bit respectively.
before, it was using sizeof(f64) == sizeof(u64), but i realized - on a platform with non-standard byte sizes (highly unlikely) - this wouldn't work.
new code uses sizeof(f64) * CHAR_BIT == 64, which is much more robust of a check since it actually ensures the proper bit count directly instead of assuming things about what std::uint64_t might be (might be a struct that's actually larger than 64-bits, for instance).

added a constexpr bool f80_enabled that's true iff long double is an 80-bit floating-point type.
this may be used in the future for handling extended-precision values in fancier ways - for now unused.

added a typedef called f80 - if f80_enabled is true, it aliases long double, oterwise void.
this would be used in SFINAE overloading or constexpr if statements for said fancy handling of extended-precision values.