beginning optimization branch
during program analysis, it turns out about 11% of the program runtime for my test program was in UpdateFlagsZSP() due to how common it is in x86 instructions.
added a macro called FLAG_ACCESS_MASKING which, when nonzero, uses a different algorithm involving masking to do the same job.
it seems the compiler is a lot more willing to optimize this usage.

fixed a compile error where a comma was left out of the last commit.

changed ProcessSUB_raw() to be a template function that takes the bool as a template argument instead of a runtime argument.
the behavior is switched using constexpr if.
did the same to ProcessAND_raw() and ProcessSUB_raw().

changed project to C++ std latest (must be at least 17 for constexpr if).