optimization + bug fix
added memory access functions with suffix _szc that take a sizecode instead of a size.
my hope is for the compiler to use a jump table for these functions, thus avoiding 2-3 comparisons in the second-best solution where it uses a binary search.

added more FLAG_ACCESS_MASKING stuff for INC/DEC.

rolled up some of the bit twiddling logic for flag setting in SUB/CMP - now only needs one call to Negative().

ProcessMOV() tends to be the slowest part of executing a program (at least in my test files), so i'm trying to find ways of eliminating work from it.
the handler itself is now rolled into one line - almost certain the compiler treats it exactly the same, though.

changed CPU/FPU register wrappers to use u8 for sizecode and indexing - 8-bit is typically (ever so slightly) faster than 16/32-bit on 64-bit processors due to not having the extra prefix.
also changed the operator= on these to return reference - should compile to the same thing because of how i'm using them, but who knows, maybe it'll help the compiler out somehow.

fixed a bug in FileDescriptor::operator=(&&) where it was using the swapping idiom.
now is a proper "unidirectional move" and, instead of swapping contents, closes its file and steals the other's.
this is more desirable on something like a file, where it's rather important they get closed as quickly as possible.