added rootdir cmd option
changed the ExeDir.h header - now only contains exe_dir(), which now returns const char* instead of const std::string&.

changed code in ExeDir.cpp - windows localization now properly uses the MAX_PATH variable for the size of the buffer array to create.
windows' code is now streamlined into a form that will be used for all further cross-platforming where object-oriented shorthanding makes the process more sleek.
an static object is created in the function which holds a buffer to be filled by its ctor, which is then returned.

removed the std::string insternal - not needed with the change of header.

added default implementation for exe_dir() - just returns nullptr.
this is why the function signature in the header was changed to const char*.

Link() in main.cpp now takes a const char* called rootdir - represents the directory of the installation.
if non-null, this is used for std file lookup (stdlib and _start.o, currently).
if null, gets it from exe_dir().
if exe_dir() fails, displays an error message saying it failed.

in the case of failure (or if the user wants to use different std files for some reason), cmd option --rootdir <pathspec> can be used explicitly.