more const accessors + bug fix
added const accessor for ZMM registers (via const reference).
ZMMRegister class now has const accessor methods so that this is actually useful.

added const overloads of the CPURegister and ZMMRegister operator[] so indexing can be done on const objects.
this is also a potential optimization point if we're only reading the value, as it bypasses the creation of the wrapper object (though any smart compiler should do that anyway with optimizations enabled).

fixed a bug where the vpu debug string writer was using ZMMRegister::int64() instead of uint64().
this isn't really a problem because of how C++ handles negative numbers in hex mode, but might be a bit confusing if you didn't know that detail.