bug fixes
updated visual studio:
the old version was lazier about syntax in templates: if the template wasn't instantiated it wouldn't even look at the syntax to see if it was correct.
the update fixes this and catches a typo in TryGetValue() for unordered_set.

fixed a problem with the ST_Wrapper variants where it wasn't using virtual destructors, nor was it disabling polymorphic deletion.
problem solved by disabling polymorphic deletion (much more efficient than the alternative of virtual destructors since this is theoretically no-op).
to make this work, ST_Wrapper and const_ST_Wrapper both inherit from ST_Wrapper_common, which is the actual class with the protected non-virtual dtor.
ST_Wrapper_common is the impl for const_ST_Wrapper (as it was before but under a new name).
const_ST_Wrapper now has a ctor taking ST_Wrapper for conversions (previously used implicit inheritance conversions).