finished terminal io wrappers
added the TerminalInputFileWrapper and TerminalOutputFileWrapper classes.
the output version is boring, just wraps an ostream and directly requests it to perform the writes.
the input version is what (theoretically) solves the linux io bug.
it manages an istream and an internal buffer.

upon receiving a a read request:
if the buffer is empty, performs a (potentially-blocking -- for cin, will definitely be blocking) read to fill the buffer with a line of input (terminated by \n).
it then performs a non-blocking read from the current buffer contents.