fixed really old cmd args stuff
the --help option in the C# version opens the browser to the spec page.
since there's not a cross-platform way to do that in C++, it's changed to be identical to -h.

the old shorthand cmd arg option parsers included '-' as a valid.
this was nice in allowing -l-o- if for some reason you wanted to do that, but it's kinda bad if you just misspell an actual --etc long form option.
'-' is removed from shorthands and '--' is added to long forms.