bug fix
fixed a bug where incorrect type aliasing was resulting in the optimizer optimizing away assignments to the point where the flags register could have incorrect values.
this was caused by EFLAGS() and FLAGS() aliasing the u64 value via reinterpret_cast.
this breaks the strict aliasing rules of the C++ standard and was thus undefined behavior.

this was not previously an issue (and still isn't in windows).
however, it's rearing its head now and has been corrected.
this is fixed by making EFLAGS() and FLAGS() return BitfieldWrapper instead of raw references.

further updates will examine the current code base for alias violations and correct them.