added -Wall flag
added the -Wall flag and corrected everything it complained about.

the strict aliasing bug still persists in Testing/run.asm.
i'd hoped this would fix it (or show me where it was), but apparently compilers aren't very good at detecting aliasing issues.
however, i have a hunch where the issue is and will be applying a potential patch in the next commit.