sanitizer builds + finally fixed the bug
added sanitizer build configurations to the makefile: sanitize and sanitize-opt (optimizer -O3).
these are both set to und and address sanitization.
these caught a lot of little things that i doubt would actually be a problem.
none the less, it's best to have the cleanest code possible.

while this didn't fix the Testing/run.asm bug that's persisted for a while, it is fixed now.
created Testing/boom.asm as a stripped-down version of the offending run.asm.
this narrowed the problem down to an the reallocation code of the malloc() implementation.
while the assembly wasn't wrong, it did lead me then to the sys_brk implementation.
again, while that wasn't wrong either it did lead me to the actual problem: min_mem_size was never initialized.
this would thus cause seemingly-random failures from sys_brk, and thus random failures from malloc/calloc/realloc.
this trickled out to the caller (run.asm), which wasn't checking for null.
thus we get an assignment to address 0, which triggers the access violation i've been getting.

so that's fixed now...