null terminated cmd line arg array
apparently the command line args in C are guaranteed to be a null terminated array of pointers.
i find that a bit strange, since you're already given the number of args.
regardless, so as to go along with standard low-level conventions, Computer::Initialize() now does this as well.

merged an asm update that guarantees this is correct.