Executable overhaul
added Executable.h - contains the Executable class.
does everything the old std::vector<u8> hack did but has a lot more built in error checking and some fancier stuff under the hood.
rewrote everything using executables to use this instead (only changes Initialize() and Link() for the most part).

Executable has a much simpler interface for our purposes.
it holds a content() array, which Computer is interested in.
it (separately) holds additional information that's not loaded into memory for execution.
it has a function called construct() that constructs an executable from component segments, which Link() is interested in.
all in all, this means everyone is happy with a simplified interface and changes to any aspect of it will not impact the other usage locations.

this is immediately used for add more safety catches when dealing with files:
upon save(), Executable (invisibly) writes an Executable-specific header, as well as the version of CSX64.
upon load(), Executable (invisibly) checks the header to make sure the file is actually a CSX64 executable, and also checks the version it was written in to make sure it's the same version of CSX64 that's loading it.
this ensures that the binary data of any old file can't be accidentally used as a CSX64 executable (could be dangerous if e.g. FSF is set).
this also gives us a degree of version safety, which is good because CSX64 updates are frequently not backwards compatible.

added some more exception types for handling new errors from Executable::load().
TypeError - thrown when attempting to use an object of invalid type (e.g. file is not an executable).
VersionError - thrown when attempting to use an object of an incompatible version (of CSX64).

moved Version from CSX64::Computer::Version to CSX64::Version (in Utility.h).
this makes it more accessible throughout the whole project, and i intend to make it more important down the line.

apparently i added BiggerInts as a submodule to the project but never actually ported math stuff over to it.
removed all the 128-bit math functions from Utility.h/cpp and rewrote all those usages in terms of BiggerInts (much clearer).
also, apparently i got lazy and used a reference cast in some of the calls to said function casting from e.g. u64& -> i64&.
this is technically safe under strict aliasing rules (because they're signed/unsigned variants), but is bad practice.
long story short - was bad and long, now short and sweet.

removed IsBitZeroFP() because we're not using it anymore.
kind of embarrassed i ever added it in the first place.

apparently using any function from the cctype header with a character that can't be represented as unsigned char is undefined behavior.
who knew? i've certainly never seen any code that hinted at this before...
changed all usage of said functions to first cast the char to unsigned char, as cppreference recommends.
this of course works for any character that's not EOF, but we're not dealing with files, so that's not a problem.

fixed a tiny bug caught in translation of the last commit to C#:
the error message for unrecognized short options was printing using argv[i], but i can change.
fixed this by doing what the C# impl does - hold on to the arg being processed with a separate variable.