added TIMES/IF and STATIC_ASSERT + bug fixes
added the TIMES prefix - works exactly as it does in NASM.
added the IF prefix - IF expr is equivalent to TIMES !!(expr).
IF is probably less useful than TIMES.
it exists due to how easy it was to implement after TIMES was already working.

added STATIC_ASSERT (and its coresponding assemble error AssembleError::Assertion).
it takes a critical expression and if it evaluates to zero (int or floating) generates an assemble error.
it optionally takes a second operand, which is a custom string to display along with the assertion error message.
so it basically works exactly like static_assert in C++17.

got rid of some places where i was assuming floating zero is integral zero.
this isn't necessarily true, as -0.0 is encoded as 0x8000..00, which is clearly non-zero.
added helper function IsZero() to Utility.h that takes a u64 value and a bool flag denoting it it's actually a floating point format.
it wraps all the logic to say if val is (int or floating) zero, which is then used to implement all the other stuff more conveniently.

caught a few places where i just copy-pasted retarded lines of code without actually fixing/repurposing them for the differences.
for instance logical AND and OR (in Expr) weren't even using the right hand value at all.
they were effectively performing the !!left operation (and yes, they were actually performing exactly the same code).
so oviously they were blatantly wrong and are now fixed.
fortunately, all the others were ok (aside from treating floating zero as integral zero).

fixed a spot where the char string to u64 conversion in Expr was overflowing its type.
the relevant line was: `(chars[i] & 0xff) << (i * 8)`, which just needed a cast to u64 for the left operand of the shift (otherwise it's an int and is potentially shifted by more than 32).

fixed a bunch of spots where i was moving from variables that are not regenerated after each iteration of the new TIMES logic.
this is, of course, a new potential problem created by the new TIMES logic, but is easily avoided by not using move semantics on assemble args member data.

TryProcessEQU now handles the 'aleady defined' logic on its own.
it didn't need to before because the label introduction logic was ensuring the operation was safe before the EQU directive handler was even executed.
the label introduction code now only performs the `already defined` logic if it's introducing a symbol.
this makes more sense from a local standpoint and makes the code not be mysteriously and dangerously entangled.
this also easily enables TIMES <1 equ ... (which does nothing) and TIMES 1 equ ... (which is a normal EQU) to compile, but not TIMES >1 equ (which would define the same symbol >1 times).

to facilitate all the new parsing issues introduced by the new TIMES/IF directives (because there's no delimiter between the end of the TIMES/IF expression and the instruction section), the expression parsing functions are now different (and faster, in fact).
instead of taking a string and parsing all of its contents as an expression, it takes a string by const& and a pair of begin/end indices.
expression parsing starts at the begin index and takes the largest expression posible up to but not including the end index.
it additionally takes an int by reference, which receives the last character not parsed as part of the expression.
thus, this allows us to parse an expression from a larger string and know where it stopped so we can resume parsing other stuff after it.

the other expression parsing functions are rewritten in terms of this.
e.g. they pass 0,len as the index pair and ensure that the entire string was consumed (i.e. the stopping index is equal to len).