added TIMES iter index macro + bug fix
added the TIMES iteration index macro $i.
this expands to the current iteration index (args.times_i) during expression parsing.
this is present even when not using a TIMES prefix, in which case it is always 0 (first and only iteration).
i could easily disallow its usage in this context, but i don't see how it could cause any problems.
my philosophy for the TIMES prefix is that every line has an implicit TIMES 1 prefix anyway.

fixed a bug with the declare directive.
if an argument started with a quotation of any recognizable form, it locks into the string declaration logic.
however this will fail in the case of 'A'+1, which should be treated as a critical expression that evaluates to 66.
thus the handler code is changed slightly:
if it can be parsed successfully as a string, it's treated as a string.
otherwise if it can be parsed successfully as an imm, it's treated as such.
otherwise it's an error, and we give the error message for the (failed) imm case (because it's harder to mess up a string).

merged asm updates for new test cases for the above