redid incbin
i misspoke in the last commit when i said incbin functioned the same as it did in nasm.
i'd started out wanting to make it just like in nasm, but then decided to make it take 1+ files and just include all of their contents (in order).
however, i've re-changed my mind and changed it to be the actual (usual) standard form.
this is more wordy for including multiple files, but gives more control with the starting pos and max include len (optional) arguments.