major organization improvements
remember that bash script i wrote forever ago and left in linux-dev/?? well i didn't...
it was supposed to automate the process of creating makefiles for more efficient building.
it worked, but had a bunch of undefined references to asan functions despite linking to asan with -lasan.
i decided to come back to it and found out the sanitize flags need to be provided to the linker AND MUST come after all the source files (because it's a static library).

however, bash is stupid so i rewrote it in python instead.
it performs intrinsics on all source files and generates the dependencies.
as an added step, it reorders the compilation objects to optimize for parallel building (larger source files come first).
it creates final targets: release, release-san, debug, and debug-san (-san options are sanitized builds in clang++, otherwise uses g++).

added a couple of build flags - now uses -Wall, -Wextra, -Wpedantic, and -Wshadow.
extra and shadow caught a lot of things that weren't problems but could be confusing - mostly shadowing members of AssembleArgs like op.

i was under the impression that removing unused parameters (by making them nameless) would do nothing for performance, but by Jove, the benchmarks were up slightly after doing so (2s vs 2.4s).

major reordering of a bunch of stuff.
the 4300 lines of Computer instruction impl is now moved out of the header and into its own file.
looking back on it, i left everything in the header out of lazyness back when function signatures were changing frequently - it just piled up.
merged a couple of files together (headers, specifically).

removed the asm_router_WHATEVER functions.
these were only used for constructing the string->handler mapping.
they're now all replaced with lambdas inside the file that constructs the mapping object (generated by a one-off python script, so everything's guaranteed to work without copy errors).

removed most cases of #if that were used before i migrated the project to C++17.
the issue was that some of the files were using the macros but they weren't defined (apparently g++ and clang++ consider an undefined macro name to be false, rather than an error).
the marcros are replaced with constexpr variables and the #if directives with constexpr if (from C++17).
this guarantees there aren't any problems with using undefined macros in #if.
additionally, this fixes a could of unused variable warnings from the new compiler warning flags in the case where the branch is not compiled (because the compiler knows that it could have been used, and just wasn't).

removed ++ (pre/post) from ST_wrapper - floating types don't have inc/dec so i don't even know why i included them.

added std::is_trivial checks to the Computer memory access (templated) utilities.
this ensures you don't accidentally try to write e.g. a std::string's binary representation to Computer memory.

rewrote __TryProcessVEC_FCMP_etc to be a parameterized function instead of a macro like it was before.
so essentially it's now implemented the same way it is in the C# impl (this impl of VEC_FCMP came first, but is objectively worse).

changed include queue to be std::deque instead of std::list to be more space efficient (holding pointers, which are small).

IsLittleEndian()'s impl is now much better - the old impl only checked to make sure a a particular byte was in a particular place, but that's not necessarily little-endian (e.g. little-big endian, where partial word orders are reverse).
now creates a 64-bit integer and an 8 byte array and uses std::memcmp() to make sure the contents are equal.

added overloads of BinRead() and BinWrite() that take the file, a buffer, and a length.
this is better than file.read() and file.write() because these automatically set the ios::failbit on failure (e.g. if file.gcount() doesn't match read size in BinRead()).
some usage of direct file.read/write() are now replaced with these (specifically those in ObjectFile.cpp, which is renamed from OFile.cpp).

got rid of the std::list<> overload of Contains() in favor of a std::deque<> impl (list was never used - it was a potential container type i was considering for taking object files to link).

fixed a typo where tohex() was taking its (templated) argument by value - could be bad.
changed ToUpper() and ToLower() to use referential foreach loops.

changed some string parsing functions to use std::size_t instead of int - they're never negative so this isn't a problem and eliminates some signed/unsigned warnings.

fixed a couple of linux compatibility issues from dev work in windows.