better expr parsing
better expr parsing:
smarter usage of ( and ):

( only increases depth if it's the starting char or if in non-zero depth.
this allows (1)(2) to be parsed as 2 valid expressions (1) and (2) rather than as 1 invalid expression.

) now breaks out of scanning mode if at depth zero (instead of marking the expression as invalid).
this allows (12+14) to be parsed starting at index 0 or 1 rather than only 0.
this is used in the fancier function-like operator parsing described below.

changed function-like operators to start with $.
this is a new lexical scheme that joins with pre-existing $, $$, and $i.
$, $$, and $i are moved into this parsing section and are now case-invariant (like the function-like operators).

changed function-like operator parsing to allow white space between op and call expr.
this means "$trunc  (1.2)" is valid.
this was one of the reasons for prefixing function-like operators with $ - i.e. to make them stand out for users.

added an extra TryGetValue overload that takes a reference instead of a pointer (will replace some usages of having an otherwise unused pointer param).