added f80 type (experimental)
added the f80 type to emulate 80-bit floating point computations for the fpu.
currently it can cast from/to f64 and f80 (untested).
it defines overloaded operators for floating point arithmetic, but currently they do this by casting to f64, doing the op, and casting back to f80 (so no precision benefits).