added movsx 64,32
the "real" movzx and movsx instructions don't provide a 64<-32 bit option.
granted, movzx doesn't need it because writing to any 32-bit general purpose register clears the high bits anyway.
but this is a very useful (and missing) feature for movsx.
there's a movsxd instruction which does specifically this, but it's kinda stupid to have a whole separate mnemonic for it.
additionally some assemblers allow using movsxd with a source size unspecified (assumes 32-bit), which is dangerous.
so i've decided to just allow a 64<-32 bit sign extend form of movsx.

did away with the nonsense in MOVxX_settings_byte().
it was structured to be like the other instructions do, but the insides of the if's are one-liners so the whole thing just looked awkward.
additionally, it was merging the 32/64 dest sizes into one and using complicated conditional expresions to separate them, which is ugly.
new style is just to encode the arguments and pass them to a switch to decode into the mode.

merged asm updates