extended setcc
the standard intell setcc instruction only allows 8-bit destinations.
i had artificially limited it to that as well to be more like intel.
but now that i'm trying to use the instruction, i realize not having at least a 32-bit version is kinda dumb.
so all i've done is removed the artificial restriction in the interpreter and assembler.