minor safety updates (usability)
changed --end to just -- (apparently that's something of a linux standard thingy)
removed - as a valid (no-op) shorthand char to ensure e.g. --fsf is not parsed as -,f,s,f due to a typo of --fs

assembler change:
defining a label on the same line as a SEGMENT directive is now an error.
it would be ambiguous otherwise which segment the label should go in.