added debug_mem instruction
added debug_mem instruction as mode 3 of OPCode::DEBUG.
this takes a memory address and a number of bytes and prints a memory dump starting there.
this was added due to not having any other utilities for looking at memory during debugging other than moving stuff into registers and printing debug info for registers (which clobbers them and is tedious/annoying).
this was used to debug the new string formatting logic in the asm submodule.

merged asm updates.