updated movd/movq instruction
apparently the movd/movq instructions, in addition to all their other behaviors, are able to move between scalar xmm registers and cpu registers.
i had no idea this was the case until i looked at the disassembly for an ieee punning function.
so i've created a new opcode called TRANS that will hold all extended encodings for misc data transfer utilities.

the handler for movd/movq now either calls the old handler or the new handler depending on if a cu register was speciified as one of the args.
technically, movd/movq can't move between two vector registers in standard intel, but i've decided to allow it since movss/movsd allow this.
the alternative would just be an unnecessary restriction on users.

changed memory dumps to print addresses as 64-bit (this was a holdover from the C# impl which only allows up to 32-bit addresses.