added movbs instruction
added the movbs instruction (move with byte swap).
the intel instruction is called movbe (move big-endian), but in CSX64 that's ambiguous with conditional move if below or equal (cmovbe) because i allow users to omit the 'c'.
thus it's added under the slightly-different name movbs.

logic in ProcessBSWAP() is now a helper function called ByteSwap().
this is used to implement both functions.
movbs is bound to the TRANS:4 extended opcode.

added a new encoding helper for binary op encoding patterns: disallow src imm.
this is internally performed by setting ab explicit b imm sizecode of -666, but is wrapped into a helper function for readability.

changed a bunch of bitwise ops on EFLAGS to RFLAGS instead to avoid having to go through the wrapper and just access the u64& directly.